Rename RoadContextView into RoadContextDataView and recreate an empty RoadContextView that will manage switch between empty and non empty road context views
