+++ DriveKitDriverDataTimelineUI/Components/RoadContext/RoadContextDataView.xib
@@ -11,7 +11,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="RoadContextView" customModule="DriveKitDriverDataTimelineUI" customModuleProvider="target">
+        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="RoadContextDataView" customModule="DriveKitDriverDataTimelineUI" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="390" height="169"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
